% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{Is a character string giving the file name of the input data
file.}
}
\value{
A clean version of an R data frame with the NOAA data contained in filename.
}
\description{
A read data file function into a cleaned version of a data frame.
}
\details{
\code{eq_clean_data} reads a filename from the U.S. National Oceanographic
 and Atmospheric Administration (NOAA) on signicant earthquakes around the world,
 into a clean version of an R data frame. It stops if the file does not exist.
The cleaned version of the data frame has:
i) the year, month and day  converted to the Date class, after dealing with negative years (BCE)
ii) EQ_PRIMARY, TOTAL_DEATHS, YEAR, LATITUDE and LONGITUDE columns converted to numeric class.
iii) A clean LOCATION_NAME column by calling the eq_location_clean function.
}
\examples{
\dontrun{
eq_clean_data("signif.txt")
}

\dontrun{
eq_clean_data()
}

}
\references{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
