% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(NOAAdata)
}
\arguments{
\item{NOAAdata}{Is a data frame in dplyr format of NOAA data.}
}
\value{
A version of the NOAA data frame with a clean formatted \code{LOCATION_NAME}
 column
}
\description{
A function to clean the \code{LOCATION_NAME} column from the
NOAA data frame.
}
\details{
\code{eq_location_clean} cleans the \code{LOCATION_NAME} column from
the NOAA data frame by stripping out the country name (including the colon)
and converting names to title case.
}
\examples{
\dontrun{
eq_location_clean(NOAAdf)
}

\dontrun{
eq_location_clean()
}

}
\references{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
