% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "timeline_label",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  xmin = NULL, xmax = NULL, n_max = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{xmin}{(optional): minimum date for earthquakes}

\item{xmax}{(optional): maximum date for earthquakes}

\item{n_max}{An integer corresponding to the number of top earthquakes to label,
sorted by measure of magnitude.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{...}{Additional arguments}
}
\value{
A ggplot2 graphical object with labelled geom_timeline plots.
}
\description{
A geom for labelling earthquake timeline plots created using
\code{\link{geom_timeline}}.
}
\details{
\code{geom_timeline_label}  adds a vertical line to each earthquake data
point with a text label (e.g. \code{LOCATION_NAME}). There should be an option to
subset to n_max number of earthquakes,sorted by any measure of magnitude
(e.g. \code{EQ_PRIMARY}).
}
\section{Aesthetics for \code{geom_timeline}, with the required ones indicated in bold}{

\itemize{
 \item \strong{\code{x}}: \code{DATE}
 \item \strong{\code{label}}: \code{LOCATION_NAME} or any labels to be added
 \item \code{y}: \code{COUNTRY} or another factor for stratification
 \item  \code{measure}: \code{EQ_PRIMARY} or any other measure of magnitude to be passed
 \item \code{color}
 \item \code{fill}
 \item \code{size}
 \item \code{alpha}
}
}

\examples{
\dontrun{
xmin = lubridate::ymd_hm("1016-01-01",truncated = 2)
xmax = lubridate::ymd_hm("2016-01-01",truncated = 2)
NOAAdf <- eq_clean_data("signif.txt")
NOAAdfSp <- NOAAdf \%>\%
  dplyr::filter(COUNTRY \%in\% c("SPAIN", "IRAN", "BELGIUM"))
ggplot2::ggplot(data = NOAAdfSp, ggplot2::aes(x = DATE,
                            y = COUNTRY,
)) +
  geom_timeline(ggplot2::aes(size = EQ_PRIMARY,
                    fill = TOTAL_DEATHS),
                xmin = xmin, xmax = xmax) +
  geom_timeline_label(aes(label = LOCATION_NAME, measure = EQ_PRIMARY),
                      xmin = xmin, xmax = xmax, n_max = 5) +
  theme_timeline() +
  ggplot2::labs(size = "Richter scale value", fill = "# deaths")
}

}
\references{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
