% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAAdf}
\alias{NOAAdf}
\title{The NOAA Significant Earthquake Database}
\format{Variable definitions can be found here:
 \url{https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225}}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
NOAAdf
}
\description{
The Significant Earthquake Database contains information on destructive earthquakes
from 2150 B.C. to the present that meet at least one of the following criteria:
Moderate damage (approximately $1 million or more), 10 or more deaths, Magnitude 7.5
or greater, Modified Mercalli Intensity X or greater, or the earthquake generated a
tsunami.
}
\details{
Please cite this database as:
National Geophysical Data Center / World Data Service (NGDC/WDS):
Significant Earthquake Database. National Geophysical Data Center, NOAA.
doi:10.7289/V5TD9V7K
}
\keyword{datasets}
