% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{StatTimeline}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimeline
}
\description{
An accessory stat to range dates between \code{xmin} and \code{xmax}
}
\keyword{datasets}
