% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{StatTimelineLabel}
\alias{StatTimelineLabel}
\title{StatTimelineLabel}
\format{An object of class \code{StatTimelineLabel} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimelineLabel
}
\description{
This function  creates the ggproto object for
\code{\link{stat_timeline_label}} to extract data according to \code{n_max},
\code{xmin} and \code{xmax} and measure and plot the corrresponding grid grob
for \code{\link{stat_timeline}}
}
\keyword{datasets}
