% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline()
}
\description{
An accessory function to format the theme of \code{geom_timeline}
earhtquakes plots as described in the instructions
}
